vis norme is on
