+++ includes/lem_in.h
@@ -6,7 +6,7 @@
 /*   By: jwillem- <jwillem-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/03/08 03:38:57 by jwillem-          #+#    #+#             */
-/*   Updated: 2019/05/19 02:07:44 by jwillem-         ###   ########.fr       */
+/*   Updated: 2019/05/21 05:25:11 by jwillem-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,6 +17,7 @@
 # include <stdio.h>
 # include "mlx.h"
 # include <limits.h>
+# include <math.h>
 
 # define ER	"ERROR: "
 # define SECURE_MALLOC(a)	!(a) && memory_error()
@@ -25,7 +26,6 @@
 # define ROOM_Q		map->room_q
 # define C_QUANT	map->comb.quant
 # define C_STEPS	map->comb.steps
-// # define C_WAY		map->comb.way_arr
 # define C_WAY		map->comb.paths
 # define C_SUM_LEN	map->comb.sum_len
 
@@ -43,34 +43,6 @@
 # define ANT		0x8B4513
 # define INDENT		map->mlx.indent
 
-typedef struct	s_img
-{
-	void	*img_ptr;
-	int		*data;
-	int		size_l;
-	int		bpp;
-	int		endian;
-}				t_img;
-
-typedef struct	s_mlx
-{
-	void	*mlx_ptr;
-	void	*win;
-	t_img	pic;
-	int		indent;
-	int		y_len;
-	int		x_len;
-}				t_mlx;
-
-/*
-**	Algo structures
-*/
-
-typedef enum	e_sort
-{
-	not_sorted, sorted
-}				e_sort;
-
 typedef struct	s_room
 {
 	char	*name;
@@ -83,6 +55,54 @@ typedef struct	s_room
 	t_list	*links;
 	int		ant_num;
 }				t_room;
+
+typedef struct	s_img
+{
+	void	*img_ptr;
+	int		*data;
+	int		size_l;
+	int		bpp;
+	int		endian;
+}				t_img;
+
+typedef struct	s_ant
+{
+	t_room	*r_curr;
+	// t_room	*r_prev;
+	char	*name;
+	int		x;
+	int		y;
+	int		x2;
+	int		y2;
+	double	sin;
+	double	cos;
+	double	hyp;
+	int		finished;
+}				t_ant;
+
+typedef struct	s_mlx
+{
+	void	*mlx_ptr;
+	void	*win;
+	t_img	pic;
+	int		indent;
+	int		y_len;
+	int		x_len;
+
+	t_ant	*ants;
+	int		movement;
+	int		delta;
+	int		step;
+}				t_mlx;
+
+/*
+**	Algo structures
+*/
+
+typedef enum	e_sort
+{
+	not_sorted, sorted
+}				e_sort;
 
 typedef struct	s_path
 {
@@ -141,7 +161,7 @@ int		lstadd_room(e_sort time_to_sort, t_list **room_list, \
 int		check_hash(char *line, char *edge, char *counter_edge);
 
 /*
-** ALGORITM
+** ALGORITHM
 */
 
 void	way(t_map *map, int l);
@@ -215,5 +235,10 @@ void	v_start_init(int fd, t_map *map, char *line, int print);
 void	v_end_init(int fd, t_map *map, char *line, int print);
 void 	start_vis_info(int fd, t_map *map, char *line, int print);
 void	get_vis_info(int fd, t_map *map, int print);
+//////////////////////
+int     draw_ants(t_map *map);
+int		draw_map(t_map *map);
+int 	press_space(int key, t_map *map);
+void	draw_line(t_map *map, size_t i, int curr_center, t_room *linked);
 
 #endif
